1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,7 +13,8 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
+
 	if (array == NULL || action == NULL)
 		return;
